requirements.txt
html5lib==1.0.1
matplotlib==2.2.3
minecart==0.3.0

nltk==3.4
numpy==1.15.4

opencv-python==3.4.4.19
pdfminer3k==1.3.1
Pillow==5.3.0
PyPDF2==1.26.0
pytesseract==0.2.6
python-dateutil==2.7.5
six==1.11.0
Wand==0.4.5
lxml==4.2.5
cycler==0.10.0
